Revert "add group creation and updated sidebar as per the new design"
