+++ src/components/Partner/AddStudentModal.jsx
@@ -1,7 +1,8 @@
 import React, { useState } from "react";
 import { Button, Modal, TextField, Typography } from "@mui/material";
 import { Box } from "@mui/material";
-import AddIcon from "@mui/icons-material/Add";
+import AddIcon from '@mui/icons-material/Add'; 
+ 
 
 const style = {
   position: "absolute",
@@ -18,12 +19,12 @@ const style = {
 
 const TextFieldStyle = {
   margin: "10px",
-};
+}
 
 const TypographyStyle = {
-  fontSize: "22px",
-  alignItem: "center",
-};
+  fontSize:"22px",
+  alignItem:"center"
+}
 
 const AddStudentModal = ({ onToggle }) => {
   const [values, setValues] = useState({});
@@ -40,34 +41,15 @@ const AddStudentModal = ({ onToggle }) => {
           <Box sx={{ display: "flex", flexDirection: "column", gap: "1rem" }}>
             <Typography variant="h6">Add Students</Typography>
             <Box sx={{ display: "flex" }}>
-              <TextField
-                name="name"
-                label="Student Name 1"
-                sx={TextFieldStyle}
-              />
-              <TextField name="name" label="Email 1" sx={TextFieldStyle} />
+            <TextField name="name" label="Student Name 1" sx={TextFieldStyle} />
+              <TextField name="name" label="Email 1" sx={TextFieldStyle}/>
             </Box>
             <Box sx={{ display: "flex" }}>
-              <TextField
-                name="name"
-                label="Student Name 2"
-                sx={TextFieldStyle}
-              />
-              <TextField name="name" label="Email 2" sx={TextFieldStyle} />
+              <TextField name="name" label="Student Name 2" sx={TextFieldStyle} />
+              <TextField name="name" label="Email 2" sx={TextFieldStyle}/>
             </Box>
-            <Typography
-              variant="subtitle2"
-              sx={{ color: "success.main", display: "flex", TypographyStyle }}
-            >
-              <AddIcon sx={{ mr: 1 }} /> Add Another Students
-            </Typography>
-            <Button
-              variant="contained"
-              sx={{
-                bgcolor: "success.main",
-                "&:hover": { bgcolor: "success.main" },
-              }}
-            >
+            <Typography variant="subtitle2" sx={{color:"success.main",display:"flex", TypographyStyle}}><AddIcon sx={{mr:1}}/> Add Another Students</Typography>
+            <Button variant="contained" sx={{ bgcolor: "success.main",  "&:hover": { bgcolor: "success.main" }}}>
               Create a space
             </Button>
           </Box>
