+++ src/components/Partner/CreateBatch.jsx
@@ -2,46 +2,46 @@ import React from "react";
 import Box from "@mui/material/Box";
 import Typography from "@mui/material/Typography";
 import { Button } from "@mui/material";
-import spaceShipImage from "../assets/student_illustration.svg";
-import { useOutletContext } from "react-router-dom";
-import { Add } from "@mui/icons-material";
+import spaceShipImage from "./assets/student_illustration.svg";
+import { useOutletContext, useParams } from "react-router-dom";
+import AddIcon from "@mui/icons-material/Add";
 
-const CreateSpace = () => {
-  const { handleCreateSpaceToggle, createSpaceOpen } = useOutletContext();
+const CreateBatch = () => {
+  const { handleCreateBatchToggle, createBatchOpen } = useOutletContext();
 
   return (
     <>
       <Box
         sx={{
           textAlign: "center",
-          bgcolor: "#FAFAFA",
-          pt: 10,
+          mt: 16,
           width: "100%",
           // border:"1px solid red"
         }}
       >
         <img src={spaceShipImage} alt="" />
         <Typography sx={{ fontSize: "14px", ml: 2 }}>
-          Spaces help organize your student data in smaller chunks
+          Let's create the first batch to start listing students
         </Typography>
 
         <Button
           variant="contained"
           aria-label="open drawer"
-          onClick={handleCreateSpaceToggle}
+          onClick={handleCreateBatchToggle}
           edge="start"
-          startIcon={<Add />}
           // sx={{ mr: 2, marginLeft: "16px", ...(open && { display: "none" }) }}
-          sx={{
-            margin: "10px",
-            borderRadius: "8px",
-          }}
+          sx={{ m: 1 }}
         >
-          <Typography variant="subtitle2">New Space</Typography>
+          <Typography
+            variant="subtitle2"
+            sx={{ display: "flex", alignItems: "center" }}
+          >
+            <AddIcon sx={{ marginRight: "8px" }} /> Create Batch
+          </Typography>
         </Button>
       </Box>
     </>
   );
 };
 
-export default CreateSpace;
+export default CreateBatch;
