--- src/components/Partner/Group/GroupList.jsx
@@ -1,47 +1,0 @@
-import { List, ListItemButton, Typography } from "@mui/material";
-import { useFetchGroupsQuery } from "../../../store";
-import GroupItem from "./GroupItem";
-import { Add } from "@mui/icons-material";
-
-function GroupList({ space, handleCreateGroupToggle }) {
-  const { data, loading, isError } = useFetchGroupsQuery(space);
-
-  let content;
-  if (loading) {
-    content = <h1>Loading...</h1>;
-  } else if (isError) {
-    content = <p>Error fetching groups</p>;
-  } else if (!data?.length) {
-    content = (
-      <ListItemButton
-        onClick={handleCreateGroupToggle}
-        sx={{
-          color: "text.primary",
-          display: "flex",
-          alignItems: "center",
-          gap: 1,
-          pl: 5.5,
-        }}
-      >
-        <Add sx={{ color: "#6d6d6d", width: "20px", height: "20px" }} />
-        <Typography
-          flex={1}
-          sx={{
-            fontSize: "14px",
-            // fontWeight: index === selected ? 600 : 400,
-          }}
-        >
-          Add Student Group
-        </Typography>
-      </ListItemButton>
-    );
-  } else {
-    content = data.map((group) => {
-      return <GroupItem group={group} key={group.id} />;
-    });
-  }
-
-  return <List>{content}</List>;
-}
-
-export default GroupList;
