+++ src/components/Partner/Sidebar.jsx
@@ -10,11 +10,10 @@ import IconButton from "@mui/material/IconButton";
 import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
 import ChevronRightIcon from "@mui/icons-material/ChevronRight";
 import { Button } from "@mui/material";
-import SpaceList from "../Space/SpaceList";
+import SpaceList from "./SpaceList";
 import { Link, useParams } from "react-router-dom";
 import { useEffect, useState } from "react";
-import { useFetchSinglePartnerQuery } from "../../../store";
-import { Add } from "@mui/icons-material";
+import { useFetchSinglePartnerQuery } from "../../store";
 
 const DrawerHeader = styled("div")(({ theme }) => ({
   display: "flex",
@@ -28,8 +27,7 @@ const DrawerHeader = styled("div")(({ theme }) => ({
 function Sidebar(value) {
   const { partnerId } = useParams();
   const { data, isLoading, error } = useFetchSinglePartnerQuery(partnerId);
-  const { createSpaceOpen, handleCreateSpaceToggle, handleCreateGroupToggle } =
-    value.value;
+  const { createSpaceOpen, handleCreateSpaceToggle } = value.value;
   const [partner, setPartner] = useState(null);
   const theme = useTheme();
 
@@ -47,7 +45,7 @@ function Sidebar(value) {
           width: 272,
           flexShrink: 0,
           "& .MuiDrawer-paper": {
-            // bgcolor: "#FAFAFA",
+            bgcolor: "#FAFAFA",
             width: 272,
             boxSizing: "border-box",
             position: "relative",
@@ -60,60 +58,33 @@ function Sidebar(value) {
         open={true}
       >
         <DrawerHeader>
-          <Link
-            to="/"
-            style={{
-              display: "flex",
-              alignItems: "center",
-            }}
-          >
+          <Link to="/">
             <IconButton>
               <ChevronLeftIcon />
             </IconButton>
-            <Typography
-              variant="span"
-              component="span"
-              fontFamily="Noto Sans"
-              sx={{ fontSize: "14px" }}
-              fontWeight="600"
-              color="text.primary"
-            >
-              Back
-            </Typography>
           </Link>
         </DrawerHeader>
         {partner && (
           <>
+            <Typography sx={{ m: 3 }}>{partner.name}</Typography>
             <Typography
-              sx={{ ml: 2, mb: 2, fontWeight: 600, fontSize: "18px" }}
-            >
-              {partner.name}
-            </Typography>
-            <Typography
-              color="text.secondary"
               style={{
                 fontSize: "14px",
                 fontWeight: "600",
                 marginLeft: "16px",
               }}
             >
-              Partner Spaces
+              Spaces
             </Typography>
             <Button
               variant="contained"
               color="inherit"
-              sx={{ width: "240px", margin: "16px", height: "40px" }}
+              style={{ width: "240px", margin: "16px" }}
               onClick={handleCreateSpaceToggle}
-              startIcon={<Add sx={{ height: "18px", width: "18px" }} />}
             >
-              <Typography sx={{ fontSize: "14px", fontWeight: "600" }}>
-                New Space
-              </Typography>
+              <Typography style={{ fontSize: "14px" }}>New Space</Typography>
             </Button>
-            <SpaceList
-              partner={partner}
-              handleCreateGroupToggle={handleCreateGroupToggle}
-            />
+            <SpaceList partner={partner} />
           </>
         )}
       </Drawer>
