+++ src/index.css
@@ -9,8 +9,3 @@
   font-family: inherit;
   outline: none;
 }
-
-a {
-  text-decoration: none;
-  display: block;
-}
