--- src/store/apis/groupsApi.js
@@ -1,46 +1,0 @@
-import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/dist/query/react";
-import { spacesApi } from "./spacesApi";
-
-const groupsApi = createApi({
-  reducerPath: "Groups",
-  baseQuery: fetchBaseQuery({
-    baseUrl: "https://merd-api.merakilearn.org/partners",
-    prepareHeaders: (headers, { getState }) => {
-      const token = getState().auth.token;
-      if (token) {
-        headers.set("authorization", `Bearer ${token}`);
-      }
-      return headers;
-    },
-  }),
-  endpoints(builder) {
-    return {
-      fetchGroups: builder.query({
-        providesTags: ["Group"],
-        query: (space) => {
-          return {
-            url: `/${space.id}/group`,
-            method: "GET",
-          };
-        },
-      }),
-      addGroup: builder.mutation({
-        invalidatesTags: ["Group"],
-        query: (group) => {
-          const { spaceId, ...rest } = group;
-          return {
-            url: `/create/group`,
-            params: {
-              space_id: spaceId,
-            },
-            body: { ...rest },
-            method: "POST",
-          };
-        },
-      }),
-    };
-  },
-});
-
-export const { useFetchGroupsQuery, useAddGroupMutation } = groupsApi;
-export { groupsApi };
